Add monkey patch to ruby fixnums

+++ lib/tensor_stream/monkey_patches/float.rb
@@ -1,3 +1,3 @@
-class Integer
+class Float
   include TensorStream::MonkeyPatch
 end
